Add abilities, moves, and types.
The order I'm going with for types is the more-or-less canonical order
that appears in official guides and stuff.  Apparently it dates all the
way back to the RBY instruction booklet.

It's a little inconsistent on whether Grass goes before Electric or
vice-versa.  Grass first keeps the starter types together, and it seems
to be the usual convention since at least HG/SS, so it's what I'm going
with.

I maaaaaay change my mind and go with internal order later; there's
some good precedent for that, too.  We'll see.