Add Pokémon's stats and effort yields.
A handful of stuff changed effort yield from Gen 3-4.  I started ripping
with Gen 8 and worked my way back, so for a while I was thinking, like,
if effort has never changed then should I make a separate table for it?
What a pain???  But thankfully there is that tiny bit of data to justify
having effort differ by gen as well.

I'm not sure what to do with Deoxys in Gen 3.  The stat table in Emerald
just has Normal Deoxys's stats; I searched the ROM for the six bytes I'd
expect Speed Forme's base stats to be (and I remembered to put Speed in
the middle), but it's not in there.  Same with Attack Forme's stats in
FireRed.  I copied the stats over from Gen 4, but I'm not sure what to
do with the effort yield.

Bulbapedia says that Deoxys is in Normal Forme in all games in Gen 3
when you battle it in the wild and when trading.  I suspect this also
applies to link battles, record mixing secret base battles, and any
hacked-in wild or trained Deoxyses.  So it may not even be meaningful to
ask what the other Formes' effort yield is in Gen 3.