Add type matchups.
Closes #1.

The __str__ method on the TypeMatchupResult enum (as well as the
existing AbilitySlot enum) is necessary so that the CLI dump command
represents them right.  This feels like a bit of a hack and is not very
clear.  I think the dump command should probably explicitly check for
enums.

Would also be nice if the enum value names used hyphens instead of
underscores in the db and in the CSVs.