Remove Fire Pledge and Water Pledge from Silvally in Sw/Sh.
See #33.

I explained why I did this in the last commit's message, but I just
realized I only did it for US/UM and forgot to do the same for Sw/Sh.

I spent a while fussing over the id columns for pokemon_moves and
pokemon_move_lists and making sure there were no holes and I already
have to give up on that...  I guess it's for the best it happened so
fast lol