Remove Surf from Busted Form Mimikyu in Sun/Moon.
See #33.

This is an edge case I forgot.  Busted Form Mimikyu is flagged as
learning Surf in Sun/Moon... but Disguised Form isn't, and you can't
teach TMs in battle, so Mimikyu can't actually learn Surf.

I guess it's pretty arbitrary, but I figured this was closer to the
Silvally case than the Melmetal case, so I'm removing it.