Add pokemon_instances.is_current.
This flags each Pokémon's latest appearance, not counting Let's Go
unless that's its only appearance.  When I get around to PLA I'll set
that aside as well since it's also a little weird.

This was also sitting around from 2022 and some changes to the porygo_nz
repo depend on it.  I don't love it but it's fine for now.  It's easier
and faster than the subquery I was using for this purpose before.

I mostly just want to get all this stuff from 2022 squared away and get
the Sw/Sh DLC update into production so I can leave myself a clean slate
and not risk coming back to this *again* in six months with a bunch of
code lying around lol.