Update for Scarlet/Violet.
Wheeee

I truncated Basculin's identifiers from basculin-red-striped to
basculin-red, and same with blue.  (I did this in the first place with
basculin-white, but forgot to change the others.)

I think this is more consistent with Urshifu, where I left out the
"strike" because urshifu-single-strike-gigantamax was too damn much, as
well as Zacian/Zamazenta, where I had no idea what to do with "Hero of
Many Battles" and just went -hero.

Similar story with dudunsparce (left off the -segment).

Also, as with As One, I squashed all four of Embody Aspect into one
ability.